+++ prisma/migrations/20260624221928_savings/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Category" ADD COLUMN     "target" DECIMAL(12,2);